+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - [MINOR] Added `http_util.CACHABLE_STATUS_CODES`
 - [PATCH] Moved `Requester` into a folder with same path as original
 - [MINOR] Added `date_util.datetime_from_date`
+- [MINOR] Add `json_util.dumps` and `json_util.loads` to use faster json serialization
 
 ### Changed